+++ safeDelAnnotation.sh
@@ -3,21 +3,13 @@
 echo "Cynthia Iradukunda" 
 echo "S3458858"
 
-#A constant variable to store The trash Can Directory path
-
+#Create a constant variable to store the trash Can directory path. 
 FOLDER=~/.trashCan
 
-# A variable to store the trash Can directory disk usage
-
-TotalUsage=0
-
-
-# Trap the SIGINT and run the trapFunction 
-
+#run the trapFunction when it receives a SIGINT signal.
 trap trapFunction SIGINT
 
-# A function to check if the trash Can directory does not exit and create one
-
+#a function to check if the trash can folder exists. And create one if it does not exist  
 doesTrashCanExist()
 {
 	
@@ -27,33 +19,23 @@ doesTrashCanExist()
     fi
 }
 
-# A function to list the content of the trash can directory when a user enters -l option
-
+# A function to list the content of the trash can directory. It checks to see the number of files in the trashcan. If the number is different from 0, it prints the file size, file type, and file name of each file. Otherwise, it notifies the user. 
 listContent()
 {
-   # Check if the trashCan is not empty	
-   if [ "$(ls -a $FOLDER)" ];
-   then	   
-	   echo " "
-           
-	   #List the trash can directory content by adding a header at the beginning to indicate the columns name. Also, for each line, it prints the filename, filetype using substring, and file size only. 
+   if [ ! $(ls -a $FOLDER| wc -l) -eq 0 ];
+   then	   echo " "
 	   ls -l $FOLDER | awk  'BEGIN{print "File-Name " "File-Type " "File-Size(Bytes)"} NR!=1{print $9 " " substr($1, 0, 1) " " $5}' | column -t
    else 
-	   #If, it is empty, it will notify the user.
         echo "The Trash Can is empty"
    fi	
 }
 
-# A function to delete all the content in the trash can directory when a user enters -d option
+# A function to delete all the content in the trash can content. It first ask the user to ensure they want to delete all files.If they do, then it deletes them interactively. Otherwise, it does not delete. 
 deleteTrashCanContent()
 {
-  #First ask the user if they want to delete all the content. 	
   echo "Are you sure you want to delete all the content in the Trash Can?(Y/N)"
-
-  #Read user response
   read userResponse 
 
-  #First capitalize the userResponse and then check if it is YES or Y then remove all the content in the folder. Otherwise, we notify the user that the content is not deleted. 
   if [ ${userResponse^} == "Y" -o ${userResponse^} == "Yes" ]; then
 	  rm -i $FOLDER/*
 	  echo "Content deleted"
@@ -63,20 +45,17 @@ deleteTrashCanContent()
   fi
 }
 
-# A function to display totalUsage when a user enters a -t option 
-
+#A function to display the disk usage of a the trashCan directory. It firsts get the total usage in bytes and then awk command helps to only print the number only without the Total.
 displayTotalUsage()
 {
-	# The du command only prints the total usage of all files not directories. Using the awk command, it stores  only the number in bytes without the word "total"
-	totalUsage=$(du -cba  ~/.trashCan | awk '{print $1; exit}')
-        TotalUsage=$totalUsage  
+	totalUsage=$(du -cba  ~/.trashCan | awk '{print $1; exit}') 
 	echo "The total usage of the Trash Can directory is: $totalUsage bytes."
 }
 
-# Start the trash Can monitoring process
+
+# A function to start the monitoring process. It first install the mate terminal incase the user does not have it installed. Then use the mate terminal to start a new window where the monitoring will start. 
 startMonitor()
 {	
-     #Use the dpkg-query to see if the mate-terminal emulator is instaled. If mate terminal is installed, it will start the monitoring process by opening the script in a new window terminal. Otherwise it will first install it. 
     if [[ $( dpkg-query -W -f='${status}' mate-terminal 2>/dev/null | grep -c "install ok installed" ) -eq 0 ]]
     then
     echo "==============================================================================="	    
@@ -90,40 +69,47 @@ startMonitor()
 
 }
 
-
-# A function to kill the monitoring process when a user enters -k option
-
+# A function to stop the monitoring process. It first get the pid of the mate terminal. If the exit status is 0 or successful, it kills the terminal. Otherwise it alerts the user that the monitoring process has not started yet.
 killMonitorProcesses()
 {
-    # It will first find the pidof the mate-terminal
     pidof=$(pidof mate-terminal)
 
-    #It will kill the mate-terminal
-    kill $pidof
+    if [[ $? -eq 0 ]]
+    then 
+	 echo "Killing the monitor script process...."
+         kill $pidof
+    else
+	 echo "The monitor script process has not yet started. Please start it using the -m option" 
+    fi
 }
 
-
-# A function to recover a file from the trash Can directory when a user enters -r option with file
-
+#A functions run everytime the user chooses a -r option. It first checks if the user has provides the filename and then ask for it if they did not. Then it checks to see if the file exists in the trash can. If true, recover it to the working directory otherwise it alerts the user that the file is not there. 
 recoverFile()
 {
-        # First get the user's working directory
-	workingDirectory=$(pwd)
-	
-	# First check if the file provides by the user, exists. If it does not, notify the user. Otherwise, it moves the file to the user working directory.
-	if [ ! -f "$FOLDER/$OPTARG" ]; then 
-	 echo "$OPTARG file is not recognised. Check if it exists in the trash Can directory."
+      File_name=
+
+        if [[ ! -v OPTARG ]]
+	then
+           echo "Enter the name of the file from the trash can you want to recover"
+	   read file_name
+	   File_name=$file_name
+	 else
+	    File_name=$OPTARG	 
+	fi
+
+	 workingDirectory=$(pwd)
+	 if [ ! -f "$FOLDER/$File_name" ]; then 
+	     echo "$File_name file is not recognised. Check if it exists in the trash Can directory."
          else
-		 mv -t $workingDirectory "$FOLDER/$OPTARG"
-		 echo "$OPTARG file recovered"
-         fi		 
+              mv "$FOLDER/$File_name" $workingDirectory
+	      echo "$File_name file recovered"
+         fi
+       	  
 }
-
-
-# A function to delete a file or multiple files by moving them to the trash can directory when a user enters arguments only. 
+# A funtion to safeDelete the different files. If first get the users arguments and check that each of them exist and it is a file. If exist, it adds it to the trashcan otherwise it alerts the user to input the right path 
 safeDelete()
 {
-      # First check if the given arguments the user wants to delete are files.
+      
       for var in $@
       do      
 	if [ ! -f "$var" ]; then	       	
@@ -132,37 +118,32 @@ safeDelete()
 	fi     
 	
       done
-      
-      # If they are all files and the user has inputed a correct path to those files, it will move them.
-      mv -t "$FOLDER" "$File"
+
+      mv -t "$FOLDER" "$@"
        
-      #If the move comand returned an exit status of 0. It will notify the user that the file(s) were deleted and run the checkIfTotalUsageExceed function. Otherwise, it notify the user to enter the correct path. 
       if [ $? -eq 0 ]
       then 
-         echo "$File deleted successfully"
+         echo "$@ deleted successfully"
 	 checkIfTotalUsageExceed
       else
-         echo "Try again,$File not deleted. Check to see if you have provided the correct  absolute path." 
+         echo "Try again,$@ not deleted. Check to see if you have provided the correct  absolute path." 
       fi	 
 
 }
 
-# A function to be executed every time a user enters ctrl + c or SIGINT signal.
+#The function runs everytime there is a trap. It display the number of files in the trash can directory by first listing them and counting each line. 
 trapFunction()
 {	
-
- # Get the number of all files in the trash can directory and print them. 	
  numberOfFiles=$(find $FOLDER -type f | wc -l)
  echo ""
  echo "The number of files in the TrashCan folder are: $numberOfFiles"
  exit
 }
 
-# A function to be executed every time a user runs the safeDel.sh 
+# A function to be run everytime the user start the safeDel.sh. It first get the different options the user is allowed to input. Based on the arguments or options, the user provides, it runs a specific function.
 optionDrivenMenu()
 {
  
-  # Provide getopts the options we want to capture and run different functions based on those option provided. 	
   while getopts "ldtmkr:" opt; do
 
    case "${opt}" in 
@@ -174,42 +155,35 @@ optionDrivenMenu()
    r) recoverFile "$OPTARG";;
     esac
  done 
-   
- # incase the user did not input any positional argument or option it will display the menu.
  if [[ "$#" == "0" ]]; then
 	 displayMenu;
 	 return;
  
  fi 	 
 
- # Incase the user inputed arguments with no options, it will run the safeDelete function.
  if [ "$OPTIND" == "1" ]
  then 
 	 safeDelete "$@"
  fi	 
 }
 
-# A function to display the menu. 
-
+# A function to be run in case the user does not provide an option. It first takes different options and based on the user option, it run a specific functions.  
 displayMenu()
 {
   echo "============================================================================"
   echo " Safe Delete Menu (Please enter a number assosiated with an option you want)"
   echo "============================================================================"
-
-  #First get the allowed options and store them in an array variable.
-  options=(-l -r -d -t -m -k Quit )
+  options=("List files" "Recover file(s)" "Delete files" "Display disk usage" "Start monitoring" "Kill monitoring process" Quit )
   
-  # Run a specific function based on the user's option choice.
   select opt in "${options[@]}"
    do 
        case $opt in
-	"-l") listContent;;
-	"-r") recoverFile;;
-	"-d") deleteTrashCanContent;;
-	"-t") displayTotalUsage;;
-	"-m") startMonitor;;
-	"-k") killMonitorProcesses;;
+	"List files") listContent;;
+	"Recover file(s)") recoverFile;;
+	"Delete files") deleteTrashCanContent;;
+	"Display disk usage") displayTotalUsage;;
+	"Start monitoring") startMonitor;;
+	"Kill monitoring process") killMonitorProcesses;;
         "Quit") break;;	 
         *)
 		echo "invalid option"
@@ -217,11 +191,11 @@ displayMenu()
 done	
 }
 
-# A function be run everytime files are moved to the trash Can
 
+# A function to run everytime the user adds a file to the trashcan and  the total usage exceeds 1000 bytes. 
 checkIfTotalUsageExceed(){
-
-    #Use the variable defined at the top to check if the total usage is greater then 1000 bytes.	
+    
+    TotalUsage=$(du -cba  ~/.trashCan | awk '{print $1; exit}')
     if [[ "$TotalUsage" -gt 1000 ]]
     then
          echo "Warning: the trash can directory exceeds 1Kbytes"
@@ -230,7 +204,7 @@ checkIfTotalUsageExceed(){
 
 }
 
-# A function to start the app by first calling the doesTrashCanExist function and then the optionDriveMenu function.
+#A function to start the app by running different functions such as optionDriveMenu and doesTrashCan exit. 
 startApp()
 {
 
@@ -238,5 +212,4 @@ doesTrashCanExist
 optionDrivenMenu "$@"
 }
 
-# Start the safeDel process.
 startApp "$@"
